refactor: change schema breakdown
