+++ packages/api/src/schema/userProfile/index.ts
@@ -1,0 +1,2 @@
+export * from "./models";
+export * from "./queries";
