+++ packages/api/src/schema/userProfile/models/UserProfile.ts
@@ -1,5 +1,4 @@
-import { schemaBuilder } from "../builder";
-import { prismaClient } from "../db";
+import { schemaBuilder } from "../../builder";
 
 schemaBuilder.prismaObject("UserProfile", {
   fields: (t) => ({
@@ -14,20 +13,3 @@ schemaBuilder.prismaObject("UserProfile", {
     owner: t.relation("user"),
   }),
 });
-
-schemaBuilder.queryField("userProfile", (t) =>
-  t.prismaField({
-    type: "UserProfile",
-    args: {
-      userId: t.arg.id({
-        required: true,
-        description: "The id of the user the profile belongs to",
-      }),
-    },
-    resolve: async (query, root, args) => {
-      return prismaClient.userProfile.findFirstOrThrow({
-        where: { userId: args.userId as string },
-      });
-    },
-  })
-);
