+++ packages/api/src/schema/userProfile/models/index.ts
@@ -1,0 +1,1 @@
+export * from "./UserProfile";
