+++ packages/api/src/schema/userProfile/queries/index.ts
@@ -1,0 +1,1 @@
+export * from "./userProfile";
