+++ packages/api/src/schema/users/index.ts
@@ -1,0 +1,3 @@
+export * from "./mutations";
+export * from "./queries";
+export * from "./models";
