+++ packages/api/src/schema/users/models/User.ts
@@ -1,0 +1,19 @@
+import { UserRegistrationStatus } from "@prisma/client";
+
+import { schemaBuilder } from "../../builder";
+
+schemaBuilder.enumType(UserRegistrationStatus, {
+  name: "UserRegistrationStatus",
+});
+
+schemaBuilder.prismaObject("User", {
+  fields: (t) => ({
+    id: t.exposeID("id"),
+    email: t.exposeString("email"),
+    joinedAt: t.expose("joinedAt", { type: "DateTime" }),
+    registrationStatus: t.expose("registrationStatus", {
+      type: UserRegistrationStatus,
+    }),
+    profile: t.relation("profile"),
+  }),
+});
