+++ packages/api/src/schema/users/mutations/index.ts
@@ -1,0 +1,1 @@
+export * from "./loginWithGoogle";
