+++ packages/api/src/schema/users/queries/users.ts
@@ -1,0 +1,11 @@
+import { schemaBuilder } from "../../builder";
+import { prismaClient } from "../../db";
+
+schemaBuilder.queryField("users", (t) =>
+  t.prismaField({
+    type: ["User"],
+    resolve: async (query) => {
+      return prismaClient.user.findMany(query);
+    },
+  })
+);
